% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{collSilur}
\alias{collSilur}
\title{Paleobiology Database collections of Silurian marine fossils}
\format{
A data frame with 8345 rows and 7 variables:
\describe{
\item{paleolng, paleolat}{Coordinates of a collection, rotated to
its palaeogeographic location with the tectonic plate model of
\href{https://www.earthbyte.org/}{GPlates}}
\item{collection_no, reference_no}{Unique identifier for the collection
and its published reference}
\item{environment}{One of 23 marine environment categories}
\item{max_ma, min_ma}{Bounds of the age estimate for a collection}
}
}
\source{
\url{https://paleobiodb.org/}
}
\usage{
collSilur
}
\description{
A dataset containing the (palaeo)coordinates and recorded marine environment
of 8,000 PBDB fossil collections from the Silurian, formatted and downloaded
from the Paleobiology Database on 24 June 2022.
}
\keyword{datasets}
